Update tests, move utils out of common
