+++ .openzeppelin/project.json
@@ -5,15 +5,13 @@
     "Operators": "Operators",
     "Managers": "Managers",
     "Settings": "Settings",
-    "Pools": "Pools",
-    "Deposits": "Deposits",
+    "Pool": "Pool",
     "Validators": "Validators",
-    "ValidatorTransfers": "ValidatorTransfers",
-    "Withdrawals": "Withdrawals",
     "Groups": "Groups",
     "Solos": "Solos",
-    "Wallet": "Wallet",
-    "Payments": "Payments"
+    "Payments": "Payments",
+    "SWDToken": "SWDToken",
+    "SWRToken": "SWRToken"
   },
   "dependencies": {},
   "name": "stakewise",
