+++ scripts/deploy.js
@@ -19,8 +19,6 @@ Loggy.silent(true);
   // Deploy all proxies
   await deployAllProxies({
     initialAdmin: process.env.INITIAL_ADMIN,
-    vrc: process.env.VRC,
-    dai: process.env.DAI,
     networkConfig,
   });
 })();
