+++ test/Settings.test.js
@@ -17,7 +17,7 @@ const {
   getNetworkConfig,
   deployLogicContracts,
 } = require('../deployments/common');
-const { removeNetworkFile } = require('./common/utils');
+const { removeNetworkFile } = require('./utils');
 
 const newValues = [
   ['minDepositUnit', ether('0.002')],
