+++ test/Validators.test.js
@@ -4,7 +4,7 @@ const {
   deployLogicContracts,
 } = require('../deployments/common');
 const { deployValidatorsProxy } = require('../deployments/validators');
-const { removeNetworkFile } = require('./common/utils');
+const { removeNetworkFile } = require('./utils');
 
 const Validators = artifacts.require('Validators');
 
@@ -12,7 +12,7 @@ const publicKey =
   '0xb07ef3635f585b5baeb057a45e7337ab5ba2b1205b43fac3a46e0add8aab242b0fb35a54373ad809405ca05c9cbf34c7';
 
 contract('Validators', ([_, pool, solos, groups, anyone]) => {
-  let networkConfig, validators, vrc, dai, settings, validatorId;
+  let networkConfig, validators;
 
   before(async () => {
     networkConfig = await getNetworkConfig();
