+++ test/access/Admins.test.js
@@ -9,7 +9,7 @@ const {
   getNetworkConfig,
   deployLogicContracts,
 } = require('../../deployments/common');
-const { removeNetworkFile } = require('../common/utils');
+const { removeNetworkFile } = require('../utils');
 
 const Admins = artifacts.require('Admins');
